Timestamp updated entities outside of the normal flow
By moving the timestamping of entities changed and their related changed
sessions and offerings into its own post flush process we will hopefully
avoid deadlocks which have frequently occurred here. I had to use the
query builder instead of just calling stampUpdate on the entities
because when a relationship is deleted in the flush (like
sessionDescription) it still attempts to save it in our postFlush event.
This causes issues so instead I'm only working with the IDs and not the
entities.