Log relationship changes for entities
We were not capturing what relationships changed for entities because
doctrine doesn't include this information.  Instead we have to capture
what collections were updated and then connect that to a change.  This
is not as specific as we would like to be, for example if a director is
added to a course the log will just say 'Ref:User' it will not indicate
if an administrator or director was added or deleted.  However baring
changes to doctrine this is as good as we can ask for at this point.

I also refactored the logger flow a bit to need less code.