Prevent users from editing their own account
Without this check it is possible for a user to add roles to their own
account.  This happens because doctrine keeps the User object in sync
between our authentication token and the user we modify in the
controller.  Since our process is to modify users and then check
authentication anyone can send a set of roles they want and the roles
will be applied before the authorization is checked.