Reduce expensive calls to class_exists and annotation reading methods
In order to read an annotation we first must build a reflection object
and that fails when we pass a bad class name.  Since we use
isIliosEntity and isIliosDto methods when checking the serializer they
get called many times and have to take care to error check their
parameters.  To reduce this cost here we scan the directory tree for
classes which contain these annotations on load and then use that list.

This should be very constant in production so we can cache these values
safely.