Stop guessing the data we will get
All of our documentation and examples send the correctly pluralized and
singular key for POST / PUT data.  Instead of trying to guess which is
used and then also allowing an un-keyed set we can just ask for the data
we want and then return a good error when it isn't there.