Remove production composer install
This test was put in to prevent issues where the AppKernel.php file and
the composer dependencies were out of sync with a dev composer
dependency in the production kernel tree.  Unfortunately it prevents
testing when a suitable release of the frontend cannot be found to match
the API version.  Since the version mismatch happens every couple of
weeks and then AppKernel issue has only happened once in 2 years we
should err on the side of useful and likely here.