Add sessions list to the sessiontypes endpoint
Allows for visibility into the links for a sessionType as well as some
modification.
It isn't possible to remove a session from this list using the
sessiontypes endpoint as it would cause an error since we have nothing
to replace it with in that Session.  Instead a nice error is returned
explaining the solution.