Strip bad html from all strings
We've been purifying HTML for fields that allow it, but not cleaning up
those fields where it isn't allowed. We handle this on the frontend by
not rendering the HTML in non-purified fields, but other API consumers
could still make a mistake and open themselves up to XSS attacks using
our data. This change prevents any harmful HTML from ending up in our
data.