Limit error messages in production environment
We were checking for 'production' here which isn't a valid environment
(it is called 'prod'). This caused us to send full error messages (but
not stack traces) when we should have sent a much more limited response.
Fix this issue and also invert the condition to make it harder to make
this mistake again.