Add query results caching and invalidation
This is a start on an idea to cache all of our query results and then
clear the entire results cache anytime something is saved. Ideally we
would only invalidate some items, but our current cache doesn't
implement tagging so we need to just dump the cache on save. This is
still a big boost in speed when updates are infrequent and clearing the
cache is fast.