Use ClockMock to speed up tests
This is a symfony phpunit bridge feature that replaces calls to sleep
(and other time functions) with a mock version that plays with time.
Instead of actually sleeping for two seconds the clock will just be
advanced which will save a few second every time each of these tests is
run.