Fix handling of empty cohort
When a program year is saved without a cohort (as it almost always is)
we need to handle null value the same way we handle no value. When
saving the frontend sends a `null` value here and I don't think we want
to allow program year cohort to be actually unset as it would be if we
allowed a null value for cohort in the validator.