Add autocomplete suggestions to search
Using the built in Completions suggester from Elasticsearch as it is the
simplest to implement. It will return matches when the prefix is the
same, but not mid-word matches. This is probably sufficient for our
autocomplete use case.

Incidentally fixed a bug where session title was not being set properly.