Stop relying on liip test bundle
Since fixture extraction has been extracted the only thing we needed
from this bundle was the getUrl method which I've extracted into a
fixture.

Changed the way we work with the test client. Instead of constantly
creating it the helper methods now take the existing client as an
argument. This prevents errors when the client is called twice and
removes the container from the first call.