Copy learning material contents into curriculum index
Learning material indexing is slow and doesn't need to be done very
often (only when new material are added). It would be very slow to index
materials attached to sessions as the sessions were updated. Instead we
pre-index all the materials and modify that only when they change and
then we extract the contents from that index and copy it into the
curriculum index for each session.