Split up large PDFs into smaller chunks for indexing
When a single PDF is too big for the elasticsearch limits we can split
it up into smaller page chunks. This has processing overhead, but since
learning materials are not uploaded very frequently that is worth paying
to get as much content into the index as possible.

Since a single LM file could now be split up into many pieces the
learning material ID is no longer a valid ID, store it in its own field.