Split messages into two busses
The performance enhancing middleware for doctrine doesn't make sense
when it is called during synchronous dispatch operations and it breaks
the app as it closes the default open connection to the database and
kills all running transactions. Instead I've applied these only to index
handling which is what gets run on read.