Account for more possible CAS interactions
If a user has authenticated with CAS, but has no username we need to
load the frontend to ensure that they get a nice error message.

If a user's browser is sending requests out of order only set the
redirect cookie for the first request. This ensures that a later request
to /sw.js or some other script can't reset the redirect cookie before it
is used.