Hold onto the CAS JWT cookie for continued login
If we delete this cookie then refreshing the app with a still valid JWT
in the application storage will cause a CAS re-login. Instead we can
keep the cookie around and use it again if needed to re-authenticate the
user. It is removed when the JWT would expire or when the user
manually logs out.