Setup the inflector in the bootstrap file
Each time the kernel is created we were re-adding new rules to the
static inflector instance. In a test process where the kernel is created thousands of
times we're adding thousands of items which eventually overloads PHPs
regex parser and causes errors.

Instead of setting this when booting the kernel moving it to the
bootstrap file does the same work, but only once.