Improve asset loading
Split up all of the link assets so that each one gets only the
attributes it needs. For stylesheets including an empty sizes attributes
causes safari to fail to load CSS at all. Preloads for fonts were not
working because they were missing the as and crossorigin attributes
which are required.