Refactor serializer services
These get called a LOT during an average request and so they are a
critical path for optimization. We were doing too much extra work in
extending the symfony ObjectNormalizer which has a lot of checks to deal
with complicated edge cases. Since our Entities and DTOs are all
shallow, well structured, and annotated classes we can skip all of that
and use a faster code path. The leads to as much as a 2x speed up in
larger API requests.