Recursively normalize DTOs
DTOs can contain other DTOs (UserEvents -> UserMaterial) in that case we
want to run the normalizer on those as well in order to correctly access
properties and remove null values. This requires those cases to be
annotated correctly so I've also cleaned up other annotations that
didn't make sense in the context of DTOs and replaced the
entityCollection with a plain array (when the property is an anonymous
object), array<string> when it is an ID list, and the new array<dto>
when it needs to be recursively normalized.