Create a multi-stage build Dockerfile
In order to run PHP we need both a webserver and a PHP interpreter. Ilios also has a variety of tasks (update-frontend, consume-messages, migrate-database) that have to be run as PHP console commands on different schedules. In a multi-stage build we can create all of the containers we need to run Ilios in a single Dockerfile which is then versioned and tagged along with the source code. This version also maintains the original php-apache container and since it is last in the file it will be the output of the build.