Split data into more shards
We're going to be running on a single node, but it seems to be indicated to split large or frequently searched / indexed data into multiple shards. There is a 50GB cap referenced in several places. Our curren LM index is running in the hundreds of GB so I've added more shards to split the indexing load. Doing the same for curriculum data as well to give some room for parallel processing of index and search.