Serve web app manifest correctly
Removing the leading slash from our links was causing visitors to
/courses/742 to download the file /courses/742/manifest.webmanifest
which would then return the index page and fail. I wasn't able to
determine the purpose of removing that slash and after removing it all
paths appear to be correct.

I also ensured that the .webmanifest file would be served compressed
correctly.

Found and correct apache sending a Content-Language: br/gz header when
those files were served. Since our static assets are never going to be
in a different language removing those fixed the issue.