Don't re-index courses when it isn't necessary
When courses are stored in the index an ingestedTime is added to each record. This is used to search for courses which do not need to be re-indexed based on the timestamp of the message. This prevents needing to re-process the same course many times when changes are made before the re-index message is received and processed.