Grant student advisors access to learning materials
Instead of blanking all materials for everyone this checks the
permissions of a user first. Students who are in the session as well as
student advisors attached to the session or course will now be able to
access all non-draft time released materials when viewing an event
through the school calendar.