Fix validation on authentication
This validation stopped working in the 5.x series of Symfony which
caused it to be impossible to save changes to an authentication record
which had a token invalidation attached to it. This changes to the
correct validation for this type of data (which can't be set by the user
and probably doesn't need to be validated) and adds tests for this
circumstance.