Extract options specifically
Instead of pulling all data out of the symfony command this extracts
each option and passes only valid options to the rollover. This ensures
that small changes to the symfony default options don't cause problems
in the test (and possibly prevents other weird bugs if there are
conflicts between defaults and our list).