Update Auth Controller for new firewall
Simplified the access checks in the auth controller to just send an
error if unauthenticated. This is simpler and more correct since no
access should come to this route unauthenticated.
The tests were working before the firewall was configured to block all
access to these routes. Now we need to modify them slightly to ensure
that access is actually blocked.