Remove unnecessary cache
Profiled this change and there is no difference between the cached and
uncached versions. This is probably because we use callbacks to build
portions of the schema only when they're needed. This use of callbacks
makes this cache unworkable with Symfony 5.4 so removing the cache makes
us ready to upgrade.