Add default learning material storage
Installing Ilios can be difficult and deciding where to store learning
materials makes it more difficult still. In developer environments we
were already defaulting this to /tmp which is way more risky then
storing things in the application root. The health check still ensures
that a proper production setup has set this value, but for initial
installs I think this may be sufficient and much easier.