Update to MySQL v8
In order to build arm compatible packages we have to switch to oracle
linux and mysql 8 for our container database. Since this version has
been stable for a while and is the default on most platforms with
support in AWS both RDS and Aurora I think this default bump to v8 is
OK. There isn't any reason 5.7 (or earlier versions) could not be used on
any install though.