Optimize CI container build
Instead of building everything 3x we can store the initial build and
then use it when testing. In order to speed this up I split the initial
build into two parts. The primary container build goes first and then
the secondary build is able to take advantage of everything that gets
cached initially.