Build amd/arm containers in different jobs
Docker cannot export/import a multi platform image so if we want to
re-use the build images in the test job we have to build them for a
single platform. This is OK as we'll still be testing both builds and
our final deploy task will build, tag, and push our official
multi-platform image.