Improve Performance of DTO Relationship Loading
By dropping back to SQL here we can query the join table directly for
manyToMany relationships on DTOs where we only want to get
the ID and don't need anything else. This is a significant performance
boost especially in cases where the DTO has a lot of related things
(like terms which can be connected to courses, sessions, objectives,
etc).