annotate AamcMethod DTO with openapi schema info.
even though nelmio/swagger picks up the schema definitions right for the
"primitive"/non-collection class members,
declaring them in the schema explicitly give us control over the order
in which they appear in the output.
that's what we want, otherwise session-types would appear before id, and
that's not what we want.