Restore symfony auto-scripts
These were removed when the bin/setup command was introduced to work
around a cache clearing issue with our docker container that wasn't
fully noted. As far as I can tell we don't have that issue anymore and
containers are able to build with these scripts in place. This is
important because the asset install is expected by many bundles so we
need to make sure it happens.