Add User Session Material Status
This API endpoint allows users to mark session learning materials as
completed making it easier to track deadlines, tasks, and educational
progress. We tie this to the session learning material instead of the LM
directly as the same LM being assigned in multiple sessions should be
individually trackable.

Using a bigint for the ID because this table may grow very large, that's
also why we're tracking the last modified value in order to be able to
run cleanup for very old items and some future point.