Prevent double IDs from breaking user sync
When a user is in the system twice with the same campus ID we can't
ignore the second entry just because it is disabled. This can cause
serious database issues when we attempt to update the non-disable
account's username or other unique information. We have to stop the sync
in this case.