Support Plural Filters on Two Endpoints
We have tests for these, but they only work because the relationship
requires no resolution for JSON API. We simple send an array to doctrine
instead of an ID and it's smart enough to convert it. For GraphQL we
needed to add a special pluralized filter so we need to handle that when
it is sent.