Disable Error Capture in CI
If this env isn't setup our config will attempt to look it up in the
database which for our code style test doesn't exist and isn't
configured. Instead of waiting for an empty and false DB reply the best
thing to do is to disable it in CI across the board because we don't
want it on here ever.