Disable provenance on built images
As of buildx 0.10 (which is included in the build-push-action 4.x
release this defaults to true and creates images with a provenance
attestation. This is an important feature for [supply chain
security](https://slsa.dev/provenance/),
however it's not working yet in AWS. Mostly this security feature allows
us to validate the containers we depend on as our container isn't in the
supply chain it is the final product this is less important. Still we'd
like to enable it when it's working so I've left comments on the change
in each place for easy reference.