Prevent Changing Root Status Through Patch Request
We prevent modifying the root status of a user via a PUT request in the
users controller, but we were not doing this same check for PATCH. This
would allow school administrators to upgrade any user, including
themselves, to have root permissions.

I also discovered the our existing tests for PATCH were not working
correctly because the rejection could come from the permissions system
and not the controller check if the user we tested with lacked
the privilege to modify a user.