Use docker tools for git context on containers
The build-push-action from docker manages git for us, in most cases this
has been fine. However, for our nightly re-builds we've been pulling off
of HEAD instead of the tag we expect. I've removed the checkout from
most places so it's clearer what is happening and setup the nightly
re-build to use the local git context and not re-checkout the code
before building.