phpstan lvl4 fix: ignore error.
nothing we can do about it here but to explicitly ignore this PHPStan
error.
the exception that we're catching here is not listed in the interface
method declaration, that's an omission further upstream.
in other words, this would have to be fixed upstream in Symfony proper.
ignoring it for now so that we can move on.