phpstan lvl4 fix: replaces value check with array index check.
peeking first before fetching the array value is better than acessing
the value by its key and THEN checking whether it exists or not.
if the key exists we know that it must have a value - we populated that
assoc array ourselves further up accordingly.
stan knows it, and so do we now.

this squishes other phpstan related errors while at it - it's the usual
pointless type hint via code comment that were griefing us again.