phpstan lvl4 fix: relationship-voter tests.
stan was objecting to the fact that we were retrieving the session-user
mock from the token mock using "magic", to declare expectations on them.
this cleans this up - instatiate the session-user mock, then pass it as
argument to the method that creates the token mock.

applies the usual cleanup while at it - adds type hints where missing,
mainly on method return types.