rm deprecated ContainerAwareInterface
this was throwing PHPStan into a frenzy whereever the $container class
member had its type indicated - whether it was via a type hint on the
member itself, or via a PHPDoc annotation on that member.
Either it was the wrong type hint, or something else was awry here.

I'm side-stepping this issue by removing the container injection
altogether - that interface is deprected and needs to go anyway before
we go up to Symfony 7.
Letting the container do its thing - in this case, injecting the
appropriate dataloader service - is the cleaner solution anyway.