load fixtures on test setup, for reasons.
there seems to be a doctrine-related dependency somewhere underneath the hood for the test
database to be there in order to make this component behave correctly under test.
i can't quite figure out where exactly.
mocking out the sub-dependencies would be possible, but having this as
an integration test rather than a unit test seems more appropriate here.
so  priming the test env to make this work is in order.