Consolidate docker compose setup
Using the overrides has proven to be unreliable and prone to errors
depending on when and where containers are built. Instead we now just
have a single development ready compose containers that shares internal
file storage.

This shared storage should fix some issues we've had in the past with
conflicts between the local files and the containers.