Improve the consume-messages container
Our messages container wasn't gracefully shutting down because we were
using the shell form of ENTRYPOINT which executes the process in bash,
but doesn't pass any signals into it. This resulted in not responding to
the docker stop command correctly.
By moving the delays into a script and using the exec form of entrypoint
we make the message consumer process the first process on the container
and it gets passed signals.

PHP requires the pcntl extension to be enabled to listen for any signals
so I've added that as well.

We were also relying on the HEALTHCHECK from the php-bash container
which checks for running fpm so our container was showing up as
unhealthy. We can just rely on the default by removing it which should
be fine because it's the only process on the image.

Lastly by using the exec form of the entrypoint we can pass arguments
into the container which are read by the consume messages command and
remove the dev containers and just send additional logging as a
parameter.