Comment out settings in dev environment
Symfony loads this files in an order that means settings in a the
.env.local file will be overwritten by settings in the .env.dev file.
This is almost never what you would expect and leads to needing to
specify some things in a .env.dev.local file that could otherwise be
consolidated in a single .env.local file. By commenting out all of these
vars we avoid this issue.