Sign redirect requests for CAS
We were allowing a cookie, which can be user defined, to redirect users
after CAS authentication. This could potentially be tampered with or set
by a bad actor.

By signing these cookies and validating them we remove this possibility
and ensure that whatever is supplied in the cookie was created by us.