Add missing lang attribute
This is always a good idea, we have it on the frontend, but didn't
include it in this HTML where almost all of our users load Ilios from.
Since this is the default loading language on the frontend I'm ok with
setting it to en, when a user changes their language this property gets
updated as well.