Ensure our cookie is valid
An old cookie in the old format, or something hokey may slip in here. We
should not only validate the hash, but ensure the URL is valid at all.
I've re-written the logic a bit to initially create the success response
and then replace it if we have a valid redirect.