Clear Messages When Updating the Index
When we run an index update it's going to index everything in the
system, we should clear out any existing messages so we don't double
index anything. This is the only way I could figure out how to clear
these without clearing something else that may show up in the future.