Skip Indexing Unchanged Synced Users
At the start of the sync process we set everyone to an unexamined state
and then flip that back after they're synchronized. This causes a lot of
churn in our search index as these users get re-indexed every time the
job is run. There isn't any need for that.

This change checks the fields that have been modified and skips indexing
users where only the examined flag has been set.