Add Middle Name to Directory
Added both a preferred and middle name option to configure both the user
sync and ldap search. When both are specified for a sync the preferred
middle name will take prescedence in the same way as first and last name
work now. There doesn't seem to be an ldap default standard for this
field so if it isn't configured we'll assume it isn't included in the
directory at all.

While processing UCSF data I found some large middle names so I bumped
up the size of this field and added a migration.