Store SQ Token in Workflow
After extensive dicussion with SQ and with the owners of this service at
UCSF we have determined this is the only path forward for scanning pull
requests from forks. The exposure risk for this token is very low,
essentially only allowing a bad actor to scan code we're already
scanning. Hopefully SQ will figure out how to use the
pull_request_target trigger at some point, but for now this is our best
path forward.