Fix Inline Type Comments
Found a code style tool and applied it to get some consistency in our
inline doc blocks.

Some of the comments were slightly wrong, needing to ale possibly be
null.

Also found an issue with ProgramYear where getCohort was supposed to
only return a cohort and never null, but that is a possibility.