Skip Re-indexing Courses
Improve performance of course indexing by skipping any course that has
been indexed already. When we add or update the index with the data in a
course we also store the ingestion time. When courses are updated and we
put a message in the async queue to be processed we record that time as
well. This allows skipping indexing a course if it has been indexed
since the message was added to the queue. When a course is being
heavily updated and the index process gets a bit behind this can be a
real time saver.