Add Tika Factory
Making the tika PHP library easy to use in services with a factory that checks that the URL exists and creates the client. If there is no URL then don't do anything, this is the same way we handle search where we don't know if a campus will have access to the service we need.