Extract Text from Materials
Using tika we can take a material and extract the text and store it in the filesystem. Unfortunately the API for the tika client requires that we pass it a file and not a string so we have to write a local version of each material to our temporary storage, extract it, and then remove it.