Replace Modified Doctrine Parameter API
The API for setParameters has changed in the latest doctrine and now
requires an ArrayCollection and Parameter objects. Seemed easier to just
use the same single invocation we use elsewhere so I replaced it.

Also dropped an extra colon at the end of each of the `to` formats. I
think this was a typo.