Add Pre-Extraction Checks
In ascending order of difficulty check that a learning material is
appropriate for extraction before attempting it. If there is an issue
with the contents of a learning material file being corrupt, as a final
stage, store the filename instead of extracted text. This will ensure
we don't attempt to keep processing the same bad file.