Validate Composer Lock File
Using this composer command we can ensure that the lock file is up to
date and working.
no-check-all: remove errors where we use `@stable` for the version
no-check-version: ignores that we store the version in our
composer.json file - this isn't recommended for unpublished packages but
it's where we track this critical data.