Update framework recipe for Symfony 7.2
The framework is now generating an APP_SECRET. We've overridden this
with ILIOS_SECRET so I kept that. Probably worth looking at to go back
to the default if the bundle is going to be messign with this.

Many options in our framework.yaml file are now the defaults in symfony
7 and were still in our config, I removed those.

The session handling config has been drastically simplified in the
recipe as well, we don't use the local session anyway, but now it aligns
with the provided recipe.