Add Opensearch Dashboard
This is setup for development. It's useful as a way to test queries and
explore the index visually. Primarily it's purpose is to create
visualizations for data, like log files, that are part of the index.
We'll probably never use it for that, but I've found it useful in
development.

I've disabled security as this is strictly for development.