Cache Our CI Container Builds
We can store the interim build layers into the Github Action Cache and
speed up some of our build steps. This isn't perfect because a lot of
our PHP images are based on the php-base, but we can't prebuild that
easily so we'll have to try this and see if it helps.