Float MySQL Docker Version
Docker images are available for v8 that stay up to date with the latest
release which is better for our containers. I also discovered that,
while it is required for a fresh install, serverVersion isn't needed
when a database already exists. Which is the case when we use the demo
DB. So I've removed it from those compose setups to make it easier to
stay in sync.