updates symfony/web-profiler-bundle recipe from v5.3 to v6.1.
the previously referenced ticket,
https://github.com/symfony/maker-bundle/issues/1252, has been resolved
and we've been using releases containing the fix for it already.
so i'm re-enabling serializer data collection in order to align our
config as close as possible as possible with the recipe blueprint.

the only_exceptions flag defaults to FALSE, no need to explicitly set
it. see https://symfony.com/doc/current/reference/configuration/framework.html#only-exceptions,

the other dropped flags are not in the blueprint anymore, and from what
i can tell, we did not customize the defaults here previously. so
accepting their removal at face value.