Add Compression for Bulk Indexing
We can compress the data we send to opensearch and pass the
Content-Encoding to open search and save a lot of transfer size. This
allows us to stay within a much smaller upload limit. This is supported
in all versions of open search, but needs to be configured in some
places. AWS enables it by default.