Add Command to Find Missing Index Values
Added this for testing, but we may use it someday. Look through all the
materials in the index and see if any are missing that we would expect.

This uses a new scrollSearch method on our open search service which is
the only way to return more than a few thousand results. Since we want
every material we have to keep going back to the server until we have
them all.