Mark Test Classes and Properties as Final
Without this PHPStan warns about calling unset on properties which could
have a setter, this isn't allowed by PHP. Marking all of these tests as
final is probably taking the defensive programming a bit too far, but we
don't really ever want to extend a test class so I'm good with it. In
places where we do extend and abstract test I marked the properties
themselves as final. This exposed some additional issues with properties
I had to clean up.