Replace Scroll With Sort
When fetching materials the scroll API is technically a good choice,
but the elastic PHP implementation uses a deprecated parameter and is
very noisy because of this deprecation. No response to this issue
upstream, so let's remove it here. This will make it much easier to
examine the logs created during indexing for any issues.