Split Failed Course Index Requests
We index courses in bulk, and mostly this is a good idea. However
sometimes a particular group of courses may cause failures or put too
much strain on the open search service. To account for this I'm catching
any exceptions in the indexing process and re-dispatching the index
requests as single courses. Since the exception is caught the message is
considered to have been handled successfully, so it won't show up as a
failure. If a single course fails we'll put that in the retry queue.