Update phpunit/phpunit Recipe
This is the new symfony flex recipe for phpunit 12, it drops the bridge
in favor of the internal deprecation handler in PHPUnit and takes
ownership of bin/phpunit which is now being used in all of our jobs.

As part of the change to for this recipe a new doctrine/deprecations
recipe was added as well. I've included it in this change as it's
related and it modifies the PHPUnit configuration as well.