Sleep Between Indexing
Order is important here for performance, we want to index every material
before we index a single course. Without the sleep it is possible for
the first courses we add to the queue to have the same available_at
timestamp as the last materials and some random number will get picked
to index in the wrong order. Sleeping ensures there is at least a full
second between the timestamps.