Improve ReIndex Performance for Courses
When a course has been indexed more recently than the queued message we
don't need to index it again. Before we were loading the data for
indexing in the handler, but this meant loading it for every course even
those that would be skipped. Instead We now pass a list of course IDs
into the indexer, check for duplicates, and re-index if needed.

This is especially important now that learning material indexing
triggers all courses to get re-indexed. So on an initial load every
course could be indexed many many times.