Split Messages into Prioritized Transports
I've created a hierarchy of high/normal/low for our messages in order to
enforce the order they are consumed in. First materials are extracted,
then indexed, and last courses are indexed with the materials. This
ensures everything gets consumed in the right order, but has the cost of
making consuming the messages more complex.