Split Indexing by Language
We have all of our text fields setup to by analized in the three main
languages of Ilios, but this means the same data is stored four times and
processed four times when it is inserted. Instead we can detect the
language of the string and only store it in the specific subfield where
it belongs.