Exit Messenger Container Gracefully
While we were trapping signals for the message consumer we weren't doing
the same for our waiters. This forced a delay when restarting containers
when either the index or db aren't available.

Augment the entrypoint here to trap signals and then send them to a
backgrounded process for each waiter. While these processes are running
we wait for them to either finish or exit.

This is way more complicated, but seems to work.