Reduce What we Store in the Curriculum Index
The _source field isn't needed for search, it's purpose it to allow us
to get data from the original document. For curriculum search we don't
use any of this data, we only use the bare minimum needed to display the
results. By only storing this data for the source we reduce the index
size significantly and make search much faster.