Improve Material DTO Build Query
Instead of joining the extra data we need for a material DTO we can use
doctrine's IDENTITY function in the DQL to get the join keys as part of
the initial query. This allows us to skip the second query entirely and
the very costly joins. Results in a substantial speedup for the API
endpoint we use to get the data for all the sessions in a course.